Added a new file containing a formalization of the Lucas-Lehmer test for primality of Mersenne numbers, and a derived rule for applying it in particular cases. This is quite a bit faster than the usual PRIME_CONV (in this special case 2^p-1), and also scales better because it doesn't rely on factorization at all. This uses as a lemma the fact that sqrt(3) is not rational, so also added a file with a trivial proof that sqrt(p) is irrational for prime p.
Added an additional 4-variable clause to FORALL_IN_GSPEC and
EXISTS_IN_GSPEC. Added some generic theorems about the order of an
element in a semigroup

        ORDER_EXISTENCE_CARD
        ORDER_EXISTENCE_FINITE
        ORDER_EXISTENCE_GEN
        ORDER_EXISTENCE_ITER

and reworked the development of "order" in Library/pocklington.ml to
use it, so the definition is now actually different (though
equivalent). Also finally moved DIVIDES_LE back into the core from
Library/prime.ml.

Tweaked the implementation of MP so that the resulting hypothesis list
exactly matches the documentation, as in the other cases pointed out
by Bill Richter.

Added a few miscellaneous, mainly Multivariate, theorems:

        BEPPO_LEVI_MONOTONE_CONVERGENCE_DECREASING_AE
        BEPPO_LEVI_MONOTONE_CONVERGENCE_INCREASING_AE
        EXISTS_UNPAIR_THM
        FORALL_UNPAIR_THM
        INTERVAL_LOWERBOUND_NONEMPTY
        INTERVAL_UPPERBOUND_NONEMPTY
        MEASURABLE_ON_SIMPLE_FUNCTION_LIMIT_INCREASING
        MONOTONE_CONVERGENCE_DECREASING_AE
        MONOTONE_CONVERGENCE_INCREASING_AE
        NEGLIGIBLE_CONVEX_INTERIOR
        OPEN_COUNTABLE_LIMIT_ELEMENTARY