Fixed bogus links to ARITH_CONV in the help files (pointed out by Michael Beeson).
Installed Moore-Penrose pseudo-inverses in place of the existing
"matrix_inv" operation. This is essentially upwards-compatible since
it is the same as the old definition when the matrix is in fact
invertible. It gives an involution and a much cleaner set of
equivalences: a matrix is invertible/symmetric/normal/PSD/PD iff its
inverse is. It also meshes nicely with the existing totalization of
the real inverse operation (e.g. in MATRIX_INV_CMUL). The definition
was based on the presentation in Giaquinta & Modica's "Linear and 
Metric Structures and Continuity", just using a couple of orthogonal
projections.

The following theorems were improved incompatibly, besides the actual
definition "matrix_inv": the hypothesis from TRANSP_MATRIX_INV was
removed and it was generalized from square matrices to arbitrary ones,
POSITIVE_DEFINITE_INV was strengthened to an equivalence, and
INVERTIBLE_MATRIX_INV was both strengthened to an equivalence and
generalized to non-square matrices. Various new theorems were added.
Complete list of new identifiers:

        COMMUTING_MATRIX_INV_COVARIANCE
        COMMUTING_MATRIX_INV_NORMAL
        COVARIANCE_MATRIX_INV
        DET_MATRIX_INV
        DIAGONAL_MATRIX_INV
        DIAGONAL_MATRIX_INV_COMPONENT
        DIAGONAL_MATRIX_INV_EXPLICIT
        DIAGONAL_MATRIX_MUL_EXPLICIT
        HAUSDIST_TRIANGLE
        IDEMPOTENT_MATRIX_TRACE_EQ_RANK
        IMAGE_MATRIX_INV
        KERNEL_MATRIX_INV
        MATRIX_INV_0
        MATRIX_INV_CMUL
        MATRIX_INV_COVARIANCE
        MATRIX_INV_COVARIANCE_LMUL
        MATRIX_INV_COVARIANCE_RMUL
        MATRIX_INV_EQ
        MATRIX_INV_EQ_0
        MATRIX_INV_INV
        MATRIX_INV_MULTIPLE_TRANP_LEFT
        MATRIX_INV_MULTIPLE_TRANP_RIGHT
        MATRIX_INV_MUL_INNER
        MATRIX_INV_MUL_OUTER
        MATRIX_INV_ORTHOGONAL_LMUL
        MATRIX_INV_ORTHOGONAL_RMUL
        MATRIX_INV_PROJECTION_IMAGE
        MATRIX_INV_PROJECTION_IMAGE_ALT
        MATRIX_INV_TRANSP
        MATRIX_INV_UNIQUE_STRONG
        MATRIX_MUL_LCANCEL
        MATRIX_MUL_RCANCEL
        MATRIX_TRANSP_MULTIPLE_INV_LEFT
        MATRIX_TRANSP_MULTIPLE_INV_RIGHT
        MATRIX_VECTOR_MUL_INV_EQ_0
        MOORE_PENROSE_PSEUDOINVERSE
        MOORE_PENROSE_PSEUDOINVERSE_UNIQUE
        NORMAL_MATRIX_INV
        POSITIVE_SEMIDEFINITE_INV
        RANK_CONJUGATE
        RANK_DIAGONAL_MATRIX
        RANK_INVERTIBLE_LMUL
        RANK_INVERTIBLE_RMUL
        RANK_MATRIX_INV
        RANK_MATRIX_INV_LMUL
        RANK_MATRIX_INV_RMUL
        SYMMETRIC_MATRIX_INV
        SYMMETRIC_MATRIX_INV_LMUL
        SYMMETRIC_MATRIX_INV_RMUL
        TRACE_MATRIX_INV_LMUL
        TRACE_MATRIX_INV_RMUL
        UNIONS_OVER_INTERS