Removed the still extant files from the old logical kernel (type.ml, term.ml and thm.ml). These were all replaced some time ago by the single file "fusion.ml", but this was initially regarded as an experimental change needing more analysis to make sure it was sound. The HOL-in-HOL / CakeML work provides all the assurance needed in this connection: see chapter 6 of Ramana Kumar's Cambridge PhD thesis "Self-complication and Self-Verification".
Also corrected source lists in the Makefile, fixed a few obselete
URLs in comments and documentation, hid a few Metis internal theorems
so they don't show up in search results, and added a line to the
zChaff->MiniSat proof translator so it compiles out of the box on
recent g++ versions.

Finally, added a mix of new theorems, some Multivariate but some
about lists and permutations, and generalized two existing theorems
ABEL_POWER_SERIES_CONTINUOUS and ABEL_LIMIT_THEOREM from the special
point 1 to any complex number; the old versions still exist but are
now called ABEL_POWER_SERIES_CONTINUOUS_1 and ABEL_LIMIT_THEOREM_1
(this is consistent with POWER_SERIES_UNIFORM_CONVERGENCE_STOLZ[_1]).
These are the new identifiers:

        ABEL_LIMIT_THEOREM_1
        ABEL_POWER_SERIES_CONTINUOUS_1
        BOUNDED_SEPARATION_1D
        HOMEOMORPHIC_SEPARATION
        HOMEOMORPHIC_SEPARATION_SPHERE
        HOMEOMORPHISM_FSIGMANESS
        HOMEOMORPHISM_GDELTANESS
        HOMEOMORPHISM_LOCAL_COMPACTNESS
        HOMEOMORPHISM_LOCAL_CONNECTEDNESS
        HOMEOMORPHISM_LOCAL_PATH_CONNECTEDNESS
        HOMEOMORPHISM_PATH_CONNECTEDNESS
        HOMEOMORPHISM_SIMPLE_CONNECTEDNESS
        LENGTH_ZIP
        PAIRWISE_APPEND
        PAIRWISE_IMPLIES
        PAIRWISE_MAP
        PAIRWISE_TRANSITIVE
        POWER_SERIES_RADIUS_OF_CONVERGENCE
        RPOW_SUB
        RPOW_SUB_ALT
        SEPARATION_BY_CLOSED_INTERMEDIATES
        SEPARATION_BY_CLOSED_INTERMEDIATES_EQ

The most interesting theorems are HOMEOMORPHIC_SEPARATION and
HOMEOMORPHIC_SEPARATION_SPHERE, which show that the properties of
separating R^n and S^n are homeomorphic invariants, with only
the assumption of boundedness in the case of R^n. (There are
already analogous theorems HOMOTOPY_EQUIVALENT_SEPARATION and
HOMOTOPY_EQUIVALENT_SEPARATION_SPHERE that only require homotopy
equivalence not homeomorphism, but they do also assume the sets
are compact.) The proof is basically the n = 2 special case of
Kuratowski's paper "Remark on an Invariance Theorem", Fundamenta
Mathematicae vol. 37 (1950), pp. 251-2.