Fixed up the Makefile for the latest OCaml and camlp5 versions (Flemming Andersen pointed out that the build failed with OCaml 4.03 and camlp5 6.16).
Added a new finite-forcing Cartesian product "finite_prod" for type
indices, to complement the existing "finite_sum" and "finite_diff" (this
was requested on the hol-info list by Abid Rauf).

Added various theorems, mainly Multivariate and in particular a
definition of "Borel measurable" and natural generalizations of preimage
and image theorems for analytic sets from continuous to Borel measurable
functions (for example ANALYTIC_BOREL_MEASURABLE_IMAGE). The actual
definition of Borel measurable is really "Baire function" (inductively
generated from continuous functions via pointwise limits) but in the
case where the domain is a Borel set this is shown to be equivalent
(BOREL_MEASURABLE_PREIMAGE_BOREL) to the more thematic preimage-based
definition.

New definitions and theorems:

        borel_measurable_CASES
        borel_measurable_INDUCT
        borel_measurable_RULES
        finite_prod_tybij
        ANALYTIC_BOREL_MEASURABLE_IMAGE
        ANALYTIC_BOREL_MEASURABLE_PREIMAGE
        ANALYTIC_LINEAR_IMAGE
        ANALYTIC_PCROSS
        ANALYTIC_PCROSS_EQ
        ANALYTIC_TRANSLATION
        BILINEAR_CONTINUOUS_ON
        BOREL_BOREL_MEASURABLE_PREIMAGE
        BOREL_INDUCT_COMPACT_ALT
        BOREL_INDUCT_OPEN_UNIONS_INTERS
        BOREL_MEASURABLE_ADD
        BOREL_MEASURABLE_BILINEAR
        BOREL_MEASURABLE_CMUL
        BOREL_MEASURABLE_COMPONENTWISE
        BOREL_MEASURABLE_COMPOSE
        BOREL_MEASURABLE_CONST
        BOREL_MEASURABLE_CONTINUOUS_COMPOSE
        BOREL_MEASURABLE_EQ
        BOREL_MEASURABLE_EXTENSION
        BOREL_MEASURABLE_IMP_MEASURABLE_ON
        BOREL_MEASURABLE_INDICATOR
        BOREL_MEASURABLE_ON_INDICATOR
        BOREL_MEASURABLE_ON_SUBSET
        BOREL_MEASURABLE_PASTECART
        BOREL_MEASURABLE_PREIMAGE_BOREL
        BOREL_MEASURABLE_SUB
        BOREL_MEASURABLE_VSUM
        CONTINUOUS_IMP_BOREL_MEASURABLE_ON
        CROSS_INTERS
        CROSS_INTERS_INTERS
        DIMINDEX_FINITE_PROD
        DIMINDEX_HAS_SIZE_FINITE_PROD
        FINITE_PROD_IMAGE
        PCROSS_INTERS
        PCROSS_INTERS_INTERS