Added miscellaneous straightforward theorems, and also ever so slightly sharpened the existing result K_SPACE_OPEN_SUBTOPOLOGY to add another disjunct "kc_space" to the existing "hausdorff_space" and "regular_space" hypotheses. Of course the "hausdorff_space" is now redundant as it implies KC, but is kept for convenience.
While a somewhat obscure refinement, I like the fact that it seems to be
getting close to best possible of its type. It already fails for the common
"weak Hausdorff" assumption a la McCord (every continuous image of a
compact Hausdorff space in it is closed): Q* x Q* (where Q* = one-point
compactification of the rationals) is WH and compact so trivially a
k-space, but the open subset Q* x Q is not a k-space as noted here:

  https://mathoverflow.net/questions/85912/products-with-compactly-generated-spaces

This result takes in as a special case half of Wilansky's classic result
(Theorem 5 from "Between T_1 and T_2") that a one-point compactification X*
is a KC space iff X is both a KC space and a k-space. Actually, I came up
with the proof by generalizing Wilansky's argument in an obvious way and
combining it with the main lemma (K_SPACE_SUBTOPOLOGY_OPEN) from
Weddington's "on k-spaces" paper. But this is very well-trodden ground so
I suppose it's in the literature somewhere.

Here are the actually new theorem identifiers:

        CONTINUOUS_MAP_EQ_INTERIOR_PREIMAGE_SUBSET
        CONTINUOUS_MAP_EQ_INTERIOR_PREIMAGE_SUBSET_ALT
        CONTINUOUS_MAP_INTERIOR_PREIMAGE_SUBSET
        DISCRETE_TOPOLOGY_CARTESIAN_PRODUCT
        DISCRETE_TOPOLOGY_CARTESIAN_PRODUCT_EQ
        EUCLIDEANREAL_CLOSURE_OF_IRRATIONAL
        EUCLIDEANREAL_CLOSURE_OF_RATIONAL
        HEREDITARILY_NORMAL_SPACE_RETRACTION_MAP_IMAGE
        HEREDITARILY_SUBTOPOLOGY
        HOMEOMORPHIC_HEREDITARILY_NORMAL_SPACE
        KC_LOCALLY_COMPACT_SPACE
        K_KC_SPACE_SUBTOPOLOGY
        LOCALLY_COMPACT_KC_IMP_REGULAR_SPACE
        LOCALLY_COMPACT_KC_SPACE
        LOCALLY_COMPACT_KC_SPACE_ALT
        NOT_LOCALLY_COMPACT_SPACE_IRRATIONAL
        NOT_LOCALLY_COMPACT_SPACE_IRRATIONAL_GEN
        NOT_LOCALLY_COMPACT_SPACE_RATIONAL
        NOT_LOCALLY_COMPACT_SPACE_RATIONAL_GEN
        PRODUCT_TOPOLOGY_EQ
        PROD_TOPOLOGY_EQ
        RETRACT_OF_SPACE_SUBTOPOLOGY
        RETRACT_OF_SPACE_TRANS