Added a decision procedure for the universal theory of rings, RING_RULE. This follows a pretty standard strategy (see my book "Handbook of Practical Logic and Automated Theorem Proving", pp. 388-389): use conjunctive normal form then appeal to convexity of the (Horn) theory of rings to reduce to considering Horn clauses, solving them via the equivalence with ideal membership using Groebner bases. As a HOL derived rule, this is of course sound. It's not quite complete because we are using Groebner bases over Q not Z, but in practice if the original problem doesn't have non-trivial constants in, it usually works.
The HOL Light version is a bit circuitous: first we instantiate the
generic normalizer (specifically intended for "total" operations)
to the case of rings whose carrier is an entire universe type. Then
we argue that we can always embed any ring in such a ring as a
subring (see the internal lemma RING_TOTALIZATION). While this is
a bit artificial, it allows maximum re-use of the carefully crafted
general normalization code, which would otherwise need to be
reimplemented in a more general setting. (I did make a couple of
tiny tweaks to "grobner.ml" as well: not strictly needed for this
but will be useful for later integral domain variants.) The rule
can be used directly on goals or explicitly given terms to
generate handy lemmas, e.g.

  RING_RULE
   `!a x:A.
        a IN ring_carrier r /\ x IN ring_carrier r /\
        ring_mul r a (ring_mul r x a) = a
        ==> ring_mul r a (ring_mul r (ring_mul r x (ring_mul r a x)) a) = a /\
            ring_mul r (ring_mul r x (ring_mul r a x))
                 (ring_mul r a (ring_mul r x (ring_mul r a x))) =
            ring_mul r x (ring_mul r a x)`;;

It's also permissible not to include the required carrier
membership assertions in the term given to RING_RULE, in which
case the returned theorem will include them (or those that seem
to be needed on naive grounds) as an antecedent, e.g.

  RING_RULE
   `!a b:C.
          (?c. c IN ring_carrier r /\ ring_mul r c a = b) /\
          (?d. d IN ring_carrier r /\ ring_mul r b d = a)
          ==> (ring_neg r a = ring_0 r <=> ring_0 r = b)`;;

  ...

  val it : thm =
    |- !a b.
           a IN ring_carrier r /\ b IN ring_carrier r
           ==> (?c. c IN ring_carrier r /\ ring_mul r c a = b) /\
               (?d. d IN ring_carrier r /\ ring_mul r b d = a)
           ==> (ring_neg r a = ring_0 r <=> ring_0 r = b)

Added a few more miscellaneous ring lemmas, and more systematically
defined the concept of a von Neumann regular ring, "vnregular_ring".
(The way it's stated with is maybe a bit artificial, since we're only
considering commutative rings, but it works fine and perhaps looks
more intuitive.) Here RING_RULE is actually quite handy for some of
the algebraic manipulations.

Also fixed up a few duplications and naming issues in the ring stuff,
both removing pointless theorems that are duplicates of others

        RING_DIVIDES_INTEGER_RING [duplicates INTEGER_RING_DIVIDES]

        RING_UNIT_INTEGER_RING    [duplicates INTEGER_RING_UNIT]

and eliminating a name conflict over RING_HOMOMORPHISM_RING_OF_INT,
renaming one of the theorems

        RING_HOMOMORPHISM_RING_OF_INT
        -> RING_HOMOMORPHISM_INTEGER_RING_OF_INT

and making a corresponding change for naming consistency

        RING_MONOMORPHISM_RING_OF_INT
        -> RING_MONOMORPHISM_INTEGER_RING_OF_INT

New definitions:

        singleton_ring
        vnregular_ring

and new theorems:

        BOOLEAN_IMP_VNREGULAR_RING
        FIELD_IMP_VNREGULAR_RING
        FIELD_MUL_LINV
        FIELD_MUL_RINV
        FINITE_SINGLETON_RING
        ISOMORPHIC_RING_SINGLETON_RING
        ISOMORPHIC_RING_VNREGULARITY
        RING_EPIMORPHISM_ONTO_FIELD_EXISTS
        RING_HOMOMORPHISM_DIAG
        RING_HOMOMORPHISM_DIAGONAL
        RING_HOMOMORPHISM_DIAGONAL_UNIV
        RING_HOMOMORPHISM_INTEGER_RING_OF_INT
        RING_MONOMORPHISM_DIAG
        RING_MONOMORPHISM_DIAGONAL
        RING_MONOMORPHISM_DIAGONAL_UNIV
        RING_MONOMORPHISM_INTEGER_RING_OF_INT
        RING_POW_IDEMPOTENT
        SINGLETON_RING
        TRIVIAL_RING_SINGLETON_RING
        VNREGULAR_IMP_BEZOUT_RING
        VNREGULAR_IMP_REDUCED_RING
        VNREGULAR_PRODUCT_RING
        VNREGULAR_PROD_RING
        VNREGULAR_QUOTIENT_RING
        VNREGULAR_RING
        VNREGULAR_RING_ALT
        VNREGULAR_RING_EPIMORPHIC_IMAGE
        VNREGULAR_RING_FINITELY_GENERATED_IDEALS
        VNREGULAR_RING_IDEAL_IDEMPOT
        VNREGULAR_RING_NILPOTENT
        VNREGULAR_RING_PRINCIPAL_IDEALS