Added some material about monomial orderings.
First there is some general stuff about how monomial divisibility is a
partial order, and on a *finite* set of variables any preorder extending
it (in particular a "compatible" one) is automatically a WQO, hence
wellfounded, and hence WO whenever total.

Then there is a more general monomial ordering monomial_le/monomial_lt
over any set of variables (the monomials themselves are still finitely
supported). This is simply the classic Dershowitz-Manna multiset
ordering, which has all the main properties we want. It's parametrized by
a variable ordering, which can be either a reflexive or irreflexive
relation. Whatever the parametrizing "order" is, the overall monomial
order is at least reflexive and transitive and compatible with
multiplication. If the variable order is a partial order on variables it
is also equivalent to the Huet-Oppen form, and in the case of a total
order becomes a lexicographic order. It also inherits being a partial
order, total order, wellorder and being wellfounded in general from the
variable order. It is *not* designed to preserve any equivalences given a
non-antisymmetric variable order. New definitions:

        monomial_deg
        monomial_le
        monomial_lt

and theorems:

        COMPATIBLE_MONOMIAL_ORDER
        FINITE_POLYNOMIAL_VARS
        FINITE_POLY_VARS
        FINITE_RESTRICTED_POWERSET
        FLD_MONOMIAL_DIVIDES
        FLD_MONOMIAL_LE
        MONOMIAL
        MONOMIAL_DEG_1
        MONOMIAL_DEG_EQ_0
        MONOMIAL_DEG_EQ_0_ALT
        MONOMIAL_DEG_LE
        MONOMIAL_DEG_MUL
        MONOMIAL_DEG_VAR
        MONOMIAL_DIV
        MONOMIAL_DIVIDES_1
        MONOMIAL_DIV_DIVIDES
        MONOMIAL_DIV_LMUL
        MONOMIAL_DIV_LMUL_EQ
        MONOMIAL_DIV_RMUL
        MONOMIAL_DIV_RMUL_EQ
        MONOMIAL_GE_1
        MONOMIAL_INDUCT
        MONOMIAL_LET_ANTISYM
        MONOMIAL_LET_MUL2
        MONOMIAL_LET_TRANS
        MONOMIAL_LE_ANTISYM
        MONOMIAL_LE_DIVISOR
        MONOMIAL_LE_IMPROPERLY
        MONOMIAL_LE_LMUL
        MONOMIAL_LE_LT
        MONOMIAL_LE_MONO
        MONOMIAL_LE_MUL2
        MONOMIAL_LE_POSET
        MONOMIAL_LE_PROPERLY
        MONOMIAL_LE_REFL
        MONOMIAL_LE_RMUL
        MONOMIAL_LE_TOSET
        MONOMIAL_LE_TRANS
        MONOMIAL_LE_VARS
        MONOMIAL_LE_VARS_EQ
        MONOMIAL_LTE_ANTISYM
        MONOMIAL_LTE_MUL2
        MONOMIAL_LTE_TRANS
        MONOMIAL_LT_ANTISYM
        MONOMIAL_LT_IMPROPERLY
        MONOMIAL_LT_IMP_LE
        MONOMIAL_LT_LMUL
        MONOMIAL_LT_MONO
        MONOMIAL_LT_MUL2
        MONOMIAL_LT_PROPERLY
        MONOMIAL_LT_REFL
        MONOMIAL_LT_RMUL
        MONOMIAL_LT_TOSET
        MONOMIAL_LT_TRANS
        MONOMIAL_LT_VARS
        MONOMIAL_LT_VARS_EQ
        MONOMIAL_MUL_LCANCEL
        MONOMIAL_MUL_RCANCEL
        MONOMIAL_RESTRICT
        MONOMIAL_VARS_DIVISOR
        MONOMIAL_VAR_DIVIDES
        POLY_CONST_0
        POLY_RING_EQ
        POLY_RING_EQ_IMP
        POLY_RING_PRODUCT_VAR
        POLY_RING_PRODUCT_VAR_POW
        POLY_RING_VAR_POW
        POSET_MONOMIAL_DIVIDES
        POSET_MONOMIAL_LE
        POWSER_RING_EQ
        POWSER_RING_EQ_IMP
        PROPERLY_MONO
        PROPERLY_MONOMIAL_LE
        QOSET_MONOMIAL_LE
        RING_MONOMORPHISM_POLY_POWSER
        TOSET_MONOMIAL_LE
        WF_COMPATIBLE_MONOMIAL_ORDER
        WF_MONOMIAL_LT
        WOSET_COMPATIBLE_MONOMIAL_ORDER
        WOSET_MONOMIAL_LE
        WQOSET_COMPATIBLE_MONOMIAL_ORDER

Also generalized the following existing ring lemmas to remove the
required carrier membership condition, since they work regardless:

        RING_PRODUCT_EQ_GENERAL_INVERSES
        RING_SUM_EQ_GENERAL_INVERSES

and added a missing outer universal quantifier to

        WF_SUBSET

Also added the following to the existing Examples/multiwf.ml file about
the wellfoundedness of the multiset ordering, asserting that the "well
founded part" is exactly the elements that can't start infinitely
descending chains:

        WFP_DCHAIN