Added a number of lemmas, mainly technical group theory results related to the correspondence between internal and external direct sums over an arbitrary indexing set, e.g.
  ISOMORPHIC_NORMAL_SUM_GROUP =
    |- !k G h.
         (!i. i IN k ==> h i normal_subgroup_of G) /\
         subgroup_generated G (UNIONS {h i | i IN k}) = G /\
         (!i. i IN k
              ==> h i INTER
                  group_carrier
                  (subgroup_generated G (UNIONS {h j | j IN k DELETE i})) =
                  {group_id G})
         ==> sum_group k (\i. subgroup_generated G (h i)) isomorphic_group G

Renamed SUBGROUP_GENERATED_SETOP -> SUBGROUP_GENERATED_SETMUL, which is a
more natural name (this was a survival from more general "op" terminology
which tried to be neutral between additive and multiplicative forms of
group properties).

Added "group_sum", "group_product", "ring_sum" and "ring_product" to the
basic congruence rules for the simplifier so that their bodies can be
simplified using the context.

Switched around all the theorems RING_SUM_RESTRICT, RING_SUM_SUPPORT,
RING_PRODUCT_RESTRICT and RING_PRODUCT_SUPPORT, to be more consistent with
similar clauses for "group_sum" and "group_product" and other iterated
operations.

Added a shallow "tactic" wrapper GROUP_TAC around GROUP_RULE, which is
marginally more convenient to use in interactive tactic proofs. It does
include the preprocessing step of trying to solve "one-relator" problems
by reducing "s = t ==> u = v" to "s = t <=> u = v" and trying to prove
that via the complete procedure for equivalences, an incomplete but on
occasional useful process for implications. (This is a decidable problem
so a complete procedcure is possible, but it requires quite a bit more
work than the equivalence case for not too many applications.)

New theorems:

        GROUP_EPIMORPHISM_ABELIAN_GROUP_SUM
        GROUP_EPIMORPHISM_GROUP_SUM
        GROUP_EPIMORPHISM_GROUP_SUM_EQ
        GROUP_EPIMORPHISM_GROUP_SUM_GEN
        GROUP_EPIMORPHISM_INTO_SUBGROUP_EQ
        GROUP_EPIMORPHISM_INTO_SUBGROUP_EQ_GEN
        GROUP_HOMOMORPHISM_GROUP_SUM
        GROUP_HOMOMORPHISM_GROUP_SUM_EQ
        GROUP_HOMOMORPHISM_GROUP_SUM_GEN
        GROUP_ISOMORPHISM_ABELIAN_GROUP_SUM
        GROUP_ISOMORPHISM_GROUP_SUM
        GROUP_ISOMORPHISM_GROUP_SUM_EQ
        GROUP_ISOMORPHISM_GROUP_SUM_GEN
        GROUP_MONOMORPHISM_ABELIAN_GROUP_SUM
        GROUP_MONOMORPHISM_GROUP_SUM
        GROUP_MONOMORPHISM_GROUP_SUM_EQ
        GROUP_MONOMORPHISM_GROUP_SUM_GEN
        GROUP_MONOMORPHISM_INTO_SUPERGROUP
        IN_SUBGROUP_PRODUCT
        IN_SUBGROUP_SUM
        ISOMORPHIC_ABELIAN_SUM_GROUP
        ISOMORPHIC_NORMAL_SUM_GROUP
        ISOMORPHIC_SUM_GROUP
        ISOMORPHIC_SUM_GROUP_GEN
        RESTRICTION_THM
        SUBGROUP_EPIMORPHISM_ABELIAN_GROUP_SUM
        SUBGROUP_EPIMORPHISM_GROUP_SUM
        SUBGROUP_EPIMORPHISM_GROUP_SUM_EQ
        SUBGROUP_EPIMORPHISM_GROUP_SUM_GEN
        SUBGROUP_GENERATED_SUPERSET
        SUBGROUP_ISOMORPHISM_ABELIAN_GROUP_SUM
        SUBGROUP_ISOMORPHISM_GROUP_SUM
        SUBGROUP_ISOMORPHISM_GROUP_SUM_EQ
        SUBGROUP_ISOMORPHISM_GROUP_SUM_GEN
        SUBGROUP_MONOMORPHISM_EPIMORPHISM
        WORD_NEG_SUB
        WORD_NOT_NOT
        WORD_XOR_NOT