Don't load 'compiler-libs.common' in OCaml 4.14
This patch makes `load_camlp5_topfind.ml` not load compiler-libs.common.
Without this patch, standard libraries like `Printf.printf` is unusable, raising
this error message:
```
Error: The module Printf is an alias for module Stdlib__Printf, which is
missing
```

I think my fix was included in my previous pull request, but it was
lost for some reason...

The reason why this line is necessary is as follows.
OCaml 4.14 REPL has its compiler-libs.common expunged by default, but
update_database_4.14.ml uses this library.
To resolve this problem,
(1) We need a toplevel whose compiler-libs.common isn't expunged. This
is done by `ocamlmktop -o ocaml-hol` as the updated README says.
(2) We need to tell `Topfind` that it must never load the library
even if other libraries need it in their dependency because `ocaml-hol`
already has that.

(2) is now resolved by this PR.

Patches for OCaml 4.14 are going a bit complicated, but yeah...

See also: https://github.com/ocaml/ocaml/issues/12271