Add `print_goal_hyp_max_boxes` to omit printing long hypotheses in goal
This commit adds a `print_goal_hyp_max_boxes` variable that can be used to omit printing too long hypotheses in a goal.

It is a reference to `int option`. If it is `Some k`, `k` is given to the max boxes option of formatter for hypotheses.
If it is `None`, nothing is set.

```
 # g `1 + 1 + 1 + 1 + 1 + 1 + 1 + 1 + 1 + 1 = 10 ==> 2 + 2 + 2 + 2 + 2 + 2 + 2 + 2 + 2 + 2 = 20`;;
val it : goalstack = 1 subgoal (1 total)

`1 + 1 + 1 + 1 + 1 + 1 + 1 + 1 + 1 + 1 = 10
 ==> 2 + 2 + 2 + 2 + 2 + 2 + 2 + 2 + 2 + 2 = 20`

 # e(STRIP_TAC);;
val it : goalstack = 1 subgoal (1 total)

  0 [`1 + 1 + 1 + 1 + 1 + 1 + 1 + 1 + 1 + 1 = 10`]

`2 + 2 + 2 + 2 + 2 + 2 + 2 + 2 + 2 + 2 = 20`

 # print_goal_hyp_max_boxes := Some 10;;
val it : unit = ()
 # p();;
val it : goalstack = 1 subgoal (1 total)

  0 [`1 + 1 + 1 + 1 + 1 + 1 + 1 + 1 + 1 + 1 = 10`]

`2 + 2 + 2 + 2 + 2 + 2 + 2 + 2 + 2 + 2 = 20`

 # print_goal_hyp_max_boxes := Some 5;;
val it : unit = ()
 # p();;
val it : goalstack = 1 subgoal (1 total)

  0 [`. = 10`]

`2 + 2 + 2 + 2 + 2 + 2 + 2 + 2 + 2 + 2 = 20`

 # print_goal_hyp_max_boxes := None;;
val it : unit = ()
 # p();;
val it : goalstack = 1 subgoal (1 total)

  0 [`1 + 1 + 1 + 1 + 1 + 1 + 1 + 1 + 1 + 1 = 10`]

`2 + 2 + 2 + 2 + 2 + 2 + 2 + 2 + 2 + 2 = 20`
```