Add `make hol.sh` that creates `hol.sh` running `ocaml` initialized with `hol.ml`
This match adds a new `hol.sh` target to Makefile. The generated `hol.sh` has
command lines that invoke `ocaml` initialized with `hol.ml`.
The command lines may vary according to the version of OCaml, and it is Makefile that
picks the right shell script template after inspecting the version.

Also, this patch
- Simplifies the continuous integration script to so that it simply launches `hol.sh` and checks any error.
- Updates README to recommend using `hol.sh` in Unix-like environment