Made some initial changes to support OCaml version 5. This is mainly simply removing the module names for some fully-qualified identifiers. These are exposed at the top level anyway and are in different places in newer versions of OCaml, e.g.
        Pervasives.compare -> Stdlib.compare
        Pervasives.sqrt -> Float.sqrt

The latter one is also captured as "float_sqrt" to avoid getting
overwritten by theorem names later.

Currently dynamic database updating is not supported in
version 5, and the current version of "update_database.ml"
is a dummy that reports that fact.

Also slightly tweaked some functions in the new bignum_zarith
wrapper to improve compatibility with the Num library, e.g
accepting negative exponents in the power function.