Pin camlp5 version of `make switch` to 8.02.01, fix typos in `NAME_ASSUMS_TAC` help
To fix the failure mentioned in https://github.com/jrh13/hol-light/pull/98 ,
This updates `make switch` to pin the camlp5 version to 8.02.01.
I previously mentioned that this version was not available on some machine in the thread,
but it seems it was available unless the opam version itself was problematic.

Furthermore, to check that `make switch` is working okay, the CI check of HOL Light is
updated to use `make switch` to set up the environment.

Another small orthogonal update is fixing typos in the help doc of `NAME_ASSUMS_TAC`.