Support compilation of HOL Light
This patch supports compilation of HOL Light.
If the `HOLLIGHT_USE_MODULE` environment variable is set to 1, Makefile builds `hol_lib.cmo` and makes `hol.sh` to use it.
It first generates `hol_lib_inlined.ml` which is the output of the new `inline_load.ml` script
which inlines all `loads`/`loadt`/`needs` invocations of the input file.
Then, it compiles the inlined file with ocmal compiler.

Compiling HOL Light with 'HOLLIGHT_USE_MODULE=1' extends the trusted base
of HOL Light to include the inliner script, inline_loads.ml. inline_loads.ml is
an OCaml program that receives an HOL Light proof and replaces the
loads/loadt/needs function invocations with their actual contents.
The README file newly explains that this flag should be turned of only if
having this additional trusted base is considered okay.