Add opam and META files, add '-dir' to hol.sh, other updates
This patch
- adds opam and META files.
- add '-dir' to hol.sh: 'hol.sh -dir' prints its HOLLIGHT_DIR variable. This is convenient when one wants to know where hol light is installed :)
- increases version at hol.ml to "3.0.0+" ('+' is to describe that there are updates after the official release)
- updates README to explain how to use OPAM to install HOL Light

The contents of 'opam' file is equivalent to the already published version to the OPAM repository
https://github.com/ocaml/opam-repository/blob/master/packages/hol_light/hol_light.3.0.0/opam
modulo:
- A reference to the released tar.gz
- `extra-source "META"` because the META file will now be included in hol-light with this patch
- Version number and some extra stuffs (they seem to extract those from the directory path)

META was necessary to help ocamlfind locate package 'hol_light'.
When HOL Light is installed during OPAM installation, I uploaded this file at gist.github.com and
download it; but I think it is a wiser idea to explicitly put it in this repo. :)