Add conversions of num/int/rat/real/word that evaluate expressions using Compute
This patch adds conversions of num/int/rat/real/word that use the Compute module to
evaluate expressions.
They are named `*_COMPUTE_CONV` and uses the 'weak' reduction (`Compute.WEAK_CBV_CONV`).
Additionally, an if-then-else expression is configured to be reduced only when the branch is constant.

These new conversions are different from

- `*_RED_CONV` because `*_RED_CONV` cannot
  recursively traverse subexpression (e.g., `NUM_RED_CONV` cannot reduce `` `1 + (2 + 3)` ``)
- Also different from `*_REDUCE_CONV` because `*_REDUCE_CONV` tries to
  recursively reduce every subexpression (e.g.,
  `` `if unknown_cond then 1 + 2 else 3 + 4` ``, `` `\x. x + (1 + 2)` ``).

This also makes github workflow fail if any command fails.