Fix 'defining multiple functions' error when a polymorphic type const is defined
This patch fixes `Failure "close_definition_clauses: defining multiple functions"` error
(github issue #121) that appears when
(1) the same definition has defined a constant with a polymorphic type
(2) the definition has several occurrences of the constant.

```
 # `(h:A list -> num) [] = 0 /\ h (CONS _ t) = 1 + h t`;;
<-- `h` is a variable, so the second `h` in the right clause is also typed as `A list -> num`

 # define `(h:A list -> num) [] = 0 /\ h (CONS _ t) = 1 + h t`;;
<-- makes `h` a constant

 # `(h:A list -> num) [] = 0 /\ h (CONS _ t) = 1 + h t`;;
<-- now `h` is a constant, so the second `h` in the right clause is assigned `:(?143198)list->num`

 # define `(h:A list -> num) [] = 0 /\ h (CONS _ t) = 1 + h t`;;
Warning: inventing type variables
Exception: Failure "close_definition_clauses: defining multiple functions".
```

This patch fixes this error by collecting multiple occurrences of the constant of the
left-hand side (the `h` constant in this example), and instantiates their
type variables (`:?143198` in this case) with one type var (`:A` in this case).