Add `HOLLIGHT_LOAD_PATH`
This patch updates `load_path` to read paths from `HOLLIGHT_LOAD_PATH`.
Hence, this affects `needs` and `loadt` which depends on `load_path`
when the loading path is relative.
Multiple paths can be added using an OS-specific separator, which is
':' for Unix and ';' for Windows.
Also, quotations ("/home/path with space", 'my dir') and escaping
(the backslash prefix) is supported.