Fix Bignum's power_num and pow not to truncate the base
This fixes a bug in the Zarith version of Bignum when the base inputs
of `power_num` and `pow` are fractions. Previously, they were incorrectly
truncating these to integers.
The `power_num` and `pow` are reimplemented to not rely on `Z.pow` anymore.
Their behavior on corner case (0^0) matches the power functions of Z (for
newer OCaml) and Num (for older OCaml).