Added a couple more theorems about subwords of composite terms:
        WORD_SUBWORD_ZX
        WORD_SUBWORD_ZX_TRIVIAL

and upgraded WORD_SIMPLE_SUBWORD_CONV to make use of those where
applicable as well as its existing theorems. This for example makes
the following work:

  WORD_SIMPLE_SUBWORD_CONV
   `word_subword ((word_zx:int256->int128) x) (16,16):int16`;;

  WORD_SIMPLE_SUBWORD_CONV
   `word_subword ((word_zx:int32->int64) x) (48,8):byte`;;

Type-generalized the existing theorems WORD_SUBWORD_AS_IWORD and
WORD_SUBWORD_IWORD to allow two distinct source and destination word
types; the constraint and even the proofs are unchanged. Also
generalized the type in the very definition "usimd16", which was
previously out of line with "usimd2", "usimd4" and "usimd8".