Made a small fix in WORD_SIMPLE_SUBWORD_CONV, which was failing to make the simple simplifications based on WORD_SUBWORD_TRIVIAL if the word operation was a composite term matching one of the other cases, e.g.
  WORD_SIMPLE_SUBWORD_CONV
    `(word_subword:(16)word->num#num->(16)word)
     ((word_join:byte->byte->int16) h l) (0,16)`;;

Fixed a bug in bit-blasting where the conversion from arithmetical to
word form was being missed in some cases where arithmetic is nested
inside other word operations (possibly after other normalizations that
started in word form). The fix is to use DEPTH_CONV in place of
ONCE_DEPTH_CONV for the WORDIFY_ATOM_CONV and IWORDIFY_ATOM_CONV
conversions. For example, this case formerly failed but now works:

  BITBLAST_RULE
  `!x:int64.
        let y = word_and x (word_sub (word_or x (word 0x0800800800800800))
                                     (word 0x0501501501501501)) in
        let z = word_ushr (word_and y (word 0x0800800800800800)) 11 in
        bitval(word_uge (word_subword x (0,12):int16) (word 3329)) +
        bitval(word_uge (word_subword x (12,12):int16) (word 3329)) +
        bitval(word_uge (word_subword x (24,12):int16) (word 3329)) +
        bitval(word_uge (word_subword x (36,12):int16) (word 3329)) +
        bitval(word_uge (word_subword x (48,12):int16) (word 3329)) =
        val(word_ushr (word_mul z (word 0x100100100100100)) 56)`;;

Added this and a few other new ones to Examples/bitblast.ml